requirements.txt
Django==1.5.5
South==0.8.2
Unipath==1.0
argparse==1.2.1
dj-database-url==0.2.2
django-allauth==0.14.1
djangorestframework==2.3.8
oauthlib==0.6.0
psycopg2==2.5.1
python-openid==2.2.5
requests==2.0.1
requests-oauthlib==0.4.0
wsgiref==0.1.2
